+++ CHANGELOG.md
@@ -1,10 +1,25 @@
+# 2.0.3
+- Voucher update 
+
+# 2.0.2
+- Address validation store-api 
+
+# 2.0.1
+- Option for headless mode 
+
+# 2.0.0
+- Shopware 6.5 compatibility
+
+# 1.1.6
+- Option to block cart if Avalara tax not calculated
+
 # 1.1.5
 - Prevent error when $_SERVER[REQUEST_URI] is not set
 
 # 1.1.4
 - Guests can edit invalid address
 
-# 1.1.3
+- # 1.1.3
 - Subshop config split
 
 # 1.1.2